Add files via upload
In this project, I analyzed live weather data sourced from an API and processed it using Pandas. The goal was to retrieve real-time weather information, clean and transform the data, and generate meaningful insights. By leveraging the flexibility of Pandas, I was able to effectively manipulate the raw JSON data returned by the API, normalize the nested structures, and convert the data into a clean, tabular format for further analysis.

The process included extracting key weather - temperature  for Calgary. 

Ultimately, this analysis provided a comprehensive understanding of temperature in Calgary, with the ability to track changes and make data-driven decisions based on real-time insights.